Graph fragment ui components seen upopoulated
