+++ app/src/main/java/com/farnsio/mindflow/ApplicationComponent.kt
@@ -3,7 +3,7 @@ package com.farnsio.mindflow
 import android.app.Application
 import com.farnsio.mindflow.modules.DataModule
 import com.farnsio.mindflow.ui.home.EditFragment
-import com.farnsio.mindflow.ui.slideshow.GraphFragment
+import com.farnsio.mindflow.ui.broken.Broken
 import dagger.BindsInstance
 import dagger.Component
 
@@ -11,7 +11,7 @@ import dagger.Component
 // Definition of the Application graph
 @Component(modules = [DataModule::class])
 interface ApplicationComponent {
-    fun inject(graphFragment: GraphFragment)
+    fun inject(broken: Broken)
     fun inject(editFragment: EditFragment)
 
     @Component.Builder
