+++ app/src/main/java/com/farnsio/mindflow/data/AppDatabase.kt
@@ -1,7 +1,7 @@
-package com.farnsio.mindflow.data;
+package com.farnsio.mindflow.data
 
-import androidx.room.Database;
-import androidx.room.RoomDatabase;
+import androidx.room.Database
+import androidx.room.RoomDatabase
 
 @Database(entities = [MentalHealthDbRecord::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
