+++ app/src/main/java/com/farnsio/mindflow/ui/broken/Broken.kt
@@ -1,4 +1,4 @@
-package com.farnsio.mindflow.ui.slideshow
+package com.farnsio.mindflow.ui.broken
 
 import android.os.Build
 import android.os.Bundle
@@ -11,15 +11,10 @@ import androidx.lifecycle.ViewModelProvider
 import com.farnsio.mindflow.MyApplication
 import com.farnsio.mindflow.R
 import com.farnsio.mindflow.data.AppDatabase
-import com.github.mikephil.charting.charts.LineChart
-import com.github.mikephil.charting.data.Entry
-import com.github.mikephil.charting.data.LineData
-import com.github.mikephil.charting.data.LineDataSet
-import java.time.LocalDateTime
 import javax.inject.Inject
 
 
-class GraphFragment : Fragment() {
+class Broken : Fragment() {
 
     private lateinit var slideshowViewModel: SlideshowViewModel
 
@@ -36,19 +31,19 @@ class GraphFragment : Fragment() {
         slideshowViewModel =
                 ViewModelProvider(this).get(SlideshowViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_graphs, container, false)
-        val lineChart: LineChart = root.findViewById(R.id.chart1)
+//        val lineChart: LineChart = root.findViewById(R.id.chart1)
 
 
-        var entries: List<Entry> = ArrayList()
-        for (data in appDatabase.mentalHealthStatusDao().getAll()) {
-            // turn your data into Entry objects
-            entries.plus(Entry(LocalDateTime.parse(data.dateTime).dayOfYear.toFloat(), data.energy.toFloat()))
-        }
-
-
-        val dataSet = LineDataSet(entries, "Energy")
-        val data : LineData = LineData(dataSet)
-        lineChart.data = data
+//        var entries: List<Entry> = ArrayList()
+//        for (data in appDatabase.mentalHealthStatusDao().getAll()) {
+//            // turn your data into Entry objects
+//            entries.plus(Entry(LocalDateTime.parse(data.dateTime).dayOfYear.toFloat(), data.energy.toFloat()))
+//        }
+//
+//
+//        val dataSet = LineDataSet(entries, "Energy")
+//        val data : LineData = LineData(dataSet)
+//        lineChart.data = data
         return root
     }
 }
