+++ app/src/main/java/com/farnsio/mindflow/ui/broken/SlideshowViewModel.kt
@@ -1,13 +1,13 @@
-package com.farnsio.mindflow.ui.gallery
+package com.farnsio.mindflow.ui.broken
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class SlideshowViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
