+++ app/src/main/java/com/farnsio/mindflow/ui/graphs/GraphFragment.kt
@@ -1,0 +1,29 @@
+package com.farnsio.mindflow.ui.graphs
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.farnsio.mindflow.R
+
+class GraphFragment : Fragment() {
+
+    private lateinit var graphViewModel: GraphViewModel
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        graphViewModel =
+                ViewModelProvider(this).get(GraphViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_graphs, container, false)
+//        val textView: TextView = root.findViewById(R.id.text_gallery)
+//        galleryViewModel.text.observe(viewLifecycleOwner, Observer {
+//            textView.text = it
+//        })
+        return root
+    }
+}
