+++ app/src/main/java/com/farnsio/mindflow/ui/graphs/GraphViewModel.kt
@@ -1,13 +1,13 @@
-package com.farnsio.mindflow.ui.slideshow
+package com.farnsio.mindflow.ui.graphs
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class GraphViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
